+++ lib/src/model/tag.dart
@@ -6,11 +6,11 @@ class Tag {
   /// Version of tag. eg. '2.4.0' or '1.1'
   String version;
 
-  /// The map of tags. Values are [dynamic] because it can be not only String.
+  /// The map of tags. Values are dynamic because it can be not only String.
   Map<String, dynamic> tags;
 
   @override
   String toString() {
-    return "TAG: $type v$version\n$tags";
+    return 'TAG: $type v$version\n$tags';
   }
 }